% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PHcure.R
\name{summary.PHcure}
\alias{summary.PHcure}
\title{Summary method for PHcure.object}
\usage{
\method{summary}{PHcure}(object, conf.int = c("basic","percentile"), conf.int.level = 0.95,...)
}
\arguments{
\item{object}{an object of class \code{\link{PHcure.object}}.}

\item{conf.int}{a character string indicating the method to compute bootstrapped confidence intervals: \code{"percentile"} or \code{"basic"}. By default \code{conf.int = "basic"}.}

\item{conf.int.level}{confidence level. By default \code{conf.int.level = 0.95}.}

\item{...}{ellipsis to pass extra arguments.}
}
\value{
An object of class \code{summary.PHcure}, a list including the following elements:
 \item{\code{N}}{the sample size (number of individuals).}
 \item{\code{censoring}}{the proportion of censored individuals.}
 \item{\code{K}}{the number of unique failure times.}
 \item{\code{isTies}}{a logical value, equal to \code{TRUE} in case of tied event times.}
 \item{\code{conf.int}}{a character string indicating the method used to compute the bootstrapped confidence intervals: \code{"percentile"}, \code{"basic"} or \code{"no"}. The latter is returned when the \code{\link{penPHcure}} function was called with the argument \code{inference = FALSE}.}
 \item{\code{conf.int.level}}{confidence level used to compute the bootstrapped confidence intervals.}
 \item{\code{logL}}{the value of the log-likelihood for the estimated model.}
 \item{\code{CURE}}{a matrix with one column containing the estimated regression coefficients in the incidence (cure) component. In case the function \code{\link{penPHcure}} was called with the argument \code{inference = TRUE}, two additional columns for the confidence intervals are provided.}
 \item{\code{SURV}}{a matrix where in the first column the estimated regression coefficients in the latency (survival) component. In case the function \code{\link{penPHcure}} was called with the argument \code{inference = TRUE}, two additional columns for the confidence intervals are provided.}
}
\description{
Produces a summary of a fitted PH cure model
}
\examples{
# For reproducibility
set.seed(3) 
# Un-comment the following line if you use R (>=3.6)
# RNGkind(sample.kind="Rounding") 

# Generate some data (for more details type ?penPHcure.simulate in your console)
data <- penPHcure.simulate()
 
# Fit standard cure model (without inference)
fit <- penPHcure(Surv(time = tstart,time2 = tstop,
                      event = status) ~ z.1 + z.2 + z.3 + z.4,
                 cureform = ~ x.1 + x.2 + x.3 + x.4,data = data)

# Use the summary method to see the results
summary(fit)
#
# ------------------------------------------------------
# +++   PH cure model with time-varying covariates   +++
# ------------------------------------------------------
# Sample size: 500
# Censoring proportion: 0.56
# Number of unique event times: 220
# Tied failure times: FALSE
# 
# log-likelihood:  101.85 
# 
# ------------------------------------------------------
# +++         Cure (incidence) coefficients          +++
# ------------------------------------------------------
#              Estimate
# (Intercept)  1.136690
# x.1         -0.816737
# x.2          0.087270
# x.3          0.910139
# x.4          0.058115
# 
# ------------------------------------------------------
# +++         Survival (latency) coefficients         +++
# ------------------------------------------------------
#      Estimate
# z.1  1.033331
# z.2 -0.051417
# z.3 -0.881202
# z.4 -0.070461

\dontrun{
# Fit standard cure model (with inference)
fit2 <- penPHcure(Surv(time = tstart,time2 = tstop,
                       event = status) ~ z.1 + z.2 + z.3 + z.4,
                  cureform = ~ x.1 + x.2 + x.3 + x.4,data = data,
                  inference = TRUE)
# Use the summary method to see the results
summary(fit2)
#
# ------------------------------------------------------
# +++   PH cure model with time-varying covariates   +++
# ------------------------------------------------------
# Sample size: 500
# Censoring proportion: 0.56
# Number of unique event times: 220
# Tied failure times: FALSE
# 
# log-likelihood:  101.85 
# 
# ------------------------------------------------------
# +++     Cure (incidence) coefficient estimates     +++
# +++          and 95\% confidence intervals *        +++
# ------------------------------------------------------
#              Estimate      2.5\%     97.5\%
# (Intercept)  1.136690  0.534618  1.526125
# x.1         -0.816737 -1.071404 -0.456797
# x.2          0.087270 -0.315229  0.469473
# x.3          0.910139  0.513726  1.251313
# x.4          0.058115 -0.321220  0.506324
# 
# ------------------------------------------------------
# +++    Survival (latency) coefficient estimates    +++
# +++          and 95\% confidence intervals *        +++
# ------------------------------------------------------
#      Estimate      2.5\%     97.5\%
# z.1  1.033331  0.855820  1.157263
# z.2 -0.051417 -0.171534  0.125073
# z.3 -0.881202 -1.003755 -0.680853
# z.4 -0.070461 -0.199304  0.098877
# 
# ------------------------------------------------------
# * Confidence intervals computed by the basic 
#   bootstrap method, with replications.
# ------------------------------------------------------

# By default, confidence intervals are computed by the basic bootstrap method.
#  Otherwise, the user can specify the percentile bootstrap method.
summary(fit2,conf.int = "percentile")
#
# ------------------------------------------------------
# +++   PH cure model with time-varying covariates   +++
# ------------------------------------------------------
# Sample size: 500
# Censoring proportion: 0.56
# Number of unique event times: 220
# Tied failure times: FALSE
# 
# log-likelihood:  101.85 
# 
# ------------------------------------------------------
# +++     Cure (incidence) coefficient estimates     +++
# +++          and 95\% confidence intervals *        +++
# ------------------------------------------------------
#              Estimate      2.5\%     97.5\%
# (Intercept)  1.136690  0.747255  1.738762
# x.1         -0.816737 -1.176677 -0.562070
# x.2          0.087270 -0.294933  0.489769
# x.3          0.910139  0.568964  1.306551
# x.4          0.058115 -0.390095  0.437450
# 
# ------------------------------------------------------
# +++    Survival (latency) coefficient estimates    +++
# +++          and 95\% confidence intervals *        +++
# ------------------------------------------------------
#      Estimate      2.5\%     97.5\%
# z.1  1.033331  0.909398  1.210841
# z.2 -0.051417 -0.227906  0.068701
# z.3 -0.881202 -1.081551 -0.758648
# z.4 -0.070461 -0.239798  0.058382
# 
# ------------------------------------------------------
# * Confidence intervals computed by the percentile 
#   bootstrap method, with replications.
# ------------------------------------------------------

# By default, a 95\% confidence level is used. Otherwise, the user can specify 
#  another confidence level: e.g. 90\%.
summary(fit2,conf.int.level = 0.90)
# 
# ------------------------------------------------------
# +++   PH cure model with time-varying covariates   +++
# ------------------------------------------------------
# Sample size: 500
# Censoring proportion: 0.56
# Number of unique event times: 220
# Tied failure times: FALSE
# 
# log-likelihood:  101.85 
# 
# ------------------------------------------------------
# +++     Cure (incidence) coefficient estimates     +++
# +++          and 90\% confidence intervals *        +++
# ------------------------------------------------------
#              Estimate        5\%       95\%
# (Intercept)  1.136690  0.625650  1.465492
# x.1         -0.816737 -1.043765 -0.491041
# x.2          0.087270 -0.233636  0.389681
# x.3          0.910139  0.616719  1.211806
# x.4          0.058115 -0.260685  0.364540
# 
# ------------------------------------------------------
# +++    Survival (latency) coefficient estimates    +++
# +++          and 90\% confidence intervals *        +++
# ------------------------------------------------------
#      Estimate        5\%       95\%
# z.1  1.033331  0.870338  1.123005
# z.2 -0.051417 -0.163104  0.081329
# z.3 -0.881202 -0.997319 -0.705101
# z.4 -0.070461 -0.189484  0.051128
# 
# ------------------------------------------------------
# * Confidence intervals computed by the basic 
#   bootstrap method, with replications.
# ------------------------------------------------------
}
}
